refactor: Extract common methods to base processor and reduce code duplication
- Move _save_processing_checkpoint() method to BaseProcessor class
- Remove duplicate implementation from all dataset processors
- Remove unused imports (time, os, json, Path) from processors
- Reduce code duplication by ~600 lines across the codebase
- Maintain full backward compatibility with existing functionality

This refactoring follows DRY (Don't Repeat Yourself) principle and improves
maintainability by centralizing common functionality in the base class.
