+++ processors/gigaspeech2.py
@@ -6,9 +6,7 @@ import os
 import logging
 from typing import Optional, Dict, Any, List, Iterator, Tuple
 import json
-import time
 import csv
-from pathlib import Path
 
 from processors.base_processor import BaseProcessor, NetworkError, ValidationError
 from utils.audio import get_audio_length, is_valid_audio
@@ -431,22 +429,6 @@ class GigaSpeech2Processor(BaseProcessor):
         
         return sample
 
-    def _save_processing_checkpoint(self, processed_count: int, current_index: int, processed_ids: set) -> None:
-        """
-        Save processing checkpoint.
-
-        Args:
-            processed_count: Number of processed samples
-            current_index: Current index for ID generation
-            processed_ids: Set of processed sample IDs
-        """
-        checkpoint_data = {
-            "processed_count": processed_count,
-            "current_index": current_index,
-            "processed_ids": list(processed_ids)
-        }
-
-        self.save_checkpoint(checkpoint_data)
 
     def get_dataset_info(self) -> Dict[str, Any]:
         """
